+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 -   Add frame interpolation (framerate) filter (thanks to orryverducci)
 -   Add -referer option for pulling HTTP streams ([PR 40](https://github.com/datarhei/restreamer-ui/pull/40), thanks to mdastgheib)
+-   Add a/v filter to the publication components ([#593](https://github.com/datarhei/restreamer-ui/issues/593))
 -   Mod uses placeholders for ingress setups ([#560](https://github.com/datarhei/restreamer-ui/issues/560))
 -   Mod updates npm
 -   Fix the advanced settings in the MPEG-TS publication service ([#597](https://github.com/datarhei/restreamer/issues/597), thanks to orryverducci)
@@ -11,7 +12,6 @@
 -   Fix index out-of-range warning, list ALSA devices for Raspicam video source
 -   Fix MUI warning
 -   Fix videojs skin
-
 
 ## v1.7.0 > v1.8.0